CHROMIUM: ARM64: dts: modify EEPROM's i2c address for kakadu and katsu
The eeprom of the front camera of katsu and kakadu cannot be read
normally. After testing, it is found that Sunny's eeprom is written
as A8, and Kingcome's is written as A4. But both of them are actually A0.
In order to facilitate os update, change the i2c address of eeprom of
Sunny and KingCome to A0.

BUG=b:185972351
TEST=build image and see if the node of eeprom exists

Cq-Depend: chromium:2887307
Change-Id: I719800f8b167a1329afba4fb2542671f6bfa8313
Signed-off-by: xiazhengqiao <xiazhengqiao@huaqin.corp-partner.google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2887206
Reviewed-by: Hsin-Yi Wang <hsinyi@chromium.org>
Reviewed-by: Wentao Qin <qinwentao@huaqin.corp-partner.google.com>
Reviewed-by: Shou-Chieh Hsu <shouchieh@chromium.org>
Commit-Queue: Jasmine Chen <lnishan@google.com>