BACKPORT: bpf: Add kconfig knob for disabling unpriv bpf by default
Add a kconfig knob which allows for unprivileged bpf to be disabled by default.
If set, the knob sets /proc/sys/kernel/unprivileged_bpf_disabled to value of 2.

This still allows a transition of 2 -> {0,1} through an admin. Similarly,
this also still keeps 1 -> {1} behavior intact, so that once set to permanently
disabled, it cannot be undone aside from a reboot.

We've also added extra2 with max of 2 for the procfs handler, so that an admin
still has a chance to toggle between 0 <-> 2.

Either way, as an additional alternative, applications can make use of CAP_BPF
that we added a while ago.

Backport Note:
* Documentation/admin-guide/sysctl/kernel.rst does not exist in
chromeos-4.19.
* The upstream patch adds BPF_UNPRIV_DEFAULT_OFF to kernel/bpf/Kconfig.
This Kconfig file does not exist on chromeos-4.19; BPF_UNPRIV_DEFAULT_OFF
is added to init/Kconfig along with some other BPF configs.
* bpf_stats_handler is not present in chromeos-4.19.
* "zero" is used in place of "SYSCTL_ZERO" in kernel/sysctl.c.

BUG=b:192962583
TEST=None

Signed-off-by: Daniel Borkmann <daniel@iogearbox.net>
Signed-off-by: Alexei Starovoitov <ast@kernel.org>
Link: https://lore.kernel.org/bpf/74ec548079189e4e4dffaeb42b8987bb3c852eee.1620765074.git.daniel@iogearbox.net
(cherry picked from commit 08389d888287c3823f80b0216766b71e17f0aba5)
Signed-off-by: Zubin Mithra <zsm@chromium.org>
Change-Id: I8d61314a3ee962f2c00774f9e9a03a1cea716f92
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3034589
Reviewed-by: Guenter Roeck <groeck@chromium.org>