CHROMIUM: Add LSM hooks to filter NVME dev access
This CL adds experimental mechanisms to see if we can guarantee
inability of userspace to directly write to NVME devices without
breaking other functionality that may need access to raw NVME devices
(e.g. debugd).
This is done by adding command points and hooks to prohibit `open`
syscall on specified block devices, and by filtering IOCTL on specified
character/block devices to exclude IO ioctls and have an allow list for
admin opcodes.
The idea is to prohibit `open` on all auxiliary namespace and dm devices
on top of it and enable ioctl filtering on main namespace and root nvme
chrdev.

Example of startup script change
https://paste.googleplex.com/5153434379485184

BUG=None
TEST=deploy on DUT and modify startup scrupt to activate the hooks
TEST=then try to open blocked block device and issue ioctls

Signed-off-by: Daniil Lunev <dlunev@chromium.org>
Change-Id: I7e551f37f87c69efd545702780052329eb096a6a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2909214
Reviewed-by: Marcus Boerger <mboerger@chromium.org>
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>