UPSTREAM: Bluetooth: mgmt: Fix the command returns garbage parameter value
When the Get Device Flags command fails, it returns the error status
with the parameters filled with the garbage values. Although the
parameters are not used, it is better to fill with zero than the random
values.

Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 02ce2c2c24024aade65a8d91d6a596651eaf2d0a)

BUG=none
TEST=build

Signed-off-by: Zhengping Jiang <jiangzp@google.com>