CHROMIUM: Bluetooth: Optimize the LE connection sequence
This patch allows the extra scan when connecting to LE devices to be
skipped if the following conditions are met:

1. A device doesn't use an Irk, so a new adv sample is not necessary.
2. A recent adv sample has been received in the last 2 seconds implying
that an rpa is likely to be current.

A manual inspection has demonstrated a ~300ms improvement to the
connection latency of the smart lock and tethering features.  This patch
is submitted as a chromium patch to gether additional field data through
metrics before suggesting upstream.

BUG=b:191003297
TEST=Manually validated using bluetoothctl and ConnectLE as well as
bluetooth_AdapterQuickHealth

Signed-off-by: Alain Michaud <alainm@chromium.org>
Change-Id: Icac3f49f9fe209193439e48d6109e183fd3c789e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2964734
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>