CHROMIUM: iwl7000: mvm: remove session protection on disassoc
If we somehow get disassociated while still waiting for a beacon
during connection, we can end up printing the

  No beacon heard and the session protection is over already...

message even if we aren't really quite waiting for it anymore.
Remove the time event, if it's running, when we get disassociated
and don't need to wait for beacons anymore.

Change-Id: Ie9c2bfdc30dcfff2c4dd7c393c79e3ac182840a9
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 3c3881bb12c7bf6c840607587373d059554e87ab