CHROMIUM: iwl7000: mvm: support RLC configuration command
There's an RLC configuration command in the firmware that's
going to replace the chain information in the PHY context.
Add support for this command, while leaving the PHY context
fields unset when we know the RLC command will be used.

Also add support to send only the RLC configuration command
if only the # of chains used on the PHY context changes.

Change-Id: I3f9d0071e680cab513c59b093d0827af99d41c51
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 4094690a475bba552eafd82c1423488aae2491e2