CHROMIUM: iwl7000: fix recognition of devices with CDB
The check to compare the CDB flag was incorrect, and caused CDB
devices to match NO_CDB entries.  Additionally, we were forcing NO_CDB
in the IWL_DEV_INFO() macro for hardcoded devices, which is not
correct.  Fix this logic.

Note that the logic was already incorrect before the cleanup patch, so
the cleanup carried the logic error to the new function.

Change-Id: If8625a4da98c718661a0f99fab8883297156a496
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: c2e4808aa8061c0de043bac3648026c9a525673d