CHROMIUM: iwl7000: don't use bit struct to decode CRF register
There are no guarantees that the order of the bits in a structure
declared with bit elements is in the declared order.  To avoid
problems, access the value using bit operations instead.

While at it, refactor and rename the get_crf_id() function for
readability.

Change-Id: I715b11da4e49ce9ea8b0a8ee0a8059e7055439ab
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: 3293ec8d65d8e553e30d8c82bda32029663d9837