CHROMIUM: drm/i915: Add module param to prevent HDMI resolution clamping
This patch adds i915 DRM module params which enables HACK to
overcome regression introduced in 4.19:
commit d61992565bd3 ("drm/i915: Determine DP++ type 1 DVI
adaptor presence based on VBT")

Mentioned commit changes the behavior of DP dual mode
detection and assume that HDMI port has a DVI type 1 adapter
connected, even though this is not actually the case.
In result, a HDMI output frequency is limited, reducing a
number of  the available refresh frequencies/resolutions and
therefore worsening user experience.  This is considered
regression for Chromebooks with built in HDMI output as user
is unable to use full capabilities of widescreen monitors.

The new parameter can be passed via kernel commandline:
i915.disable_dualmode_redetection=1

This fix should be treated as a HACK till we land proper
solution in upstream, but is needed as short term solution
for P0 BUG for Lars on M94.

BUG=b:203378940
TEST=Build and flash Lars recovery image.
     Connect HDMI widescreen monitor.
     ChromeOs should use full set of monitor resolutions.

Cq-Depend: chromium:3236995
Signed-off-by: Radoslaw Biernacki <rad@semihalf.com>
Signed-off-by: Radoslaw Biernacki <biernacki@google.com>
Fixes: d61992565bd3 ("drm/i915: Determine DP++ type 1 DVI adaptor presence based")
Change-Id: I6df0b3609be8ce6bf642798747835fc5edf36039
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3233137
Reviewed-by: Marcin Wojtas <mwojtas@google.com>
Commit-Queue: Marcin Wojtas <mwojtas@google.com>