CHROMIUM: iwl7000: mvm: don't crash on invalid rate w/o STA
If we get to the WARN_ONCE(..., "Got a HT rate (...)", ...)
here with a NULL sta, then we crash because mvmsta is bad
and we try to dereference it. Fix that by printing -1 as the
state if no station was given.

Change-Id: I606c746e11447fe168cf046376b70b04e278c3b4
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 9c8740b048b9c1e78ebb4e2ebe65bc1e8b5ff90d