CHROMIUM: Bluetooth: msft: Clear tracked devices on resume
Clear already tracked devices on system resume. Once the monitors are
reregistered after resume, matched devices in range will be found again.

Test performed:
Create an Advertisement Monitor and suspend the system once the
DeviceFound event is triggered for that monitor. Verify that the
DeviceLost event is triggered for that device on resume.

NOTE: This patch is required to fix the regression b:211793253 on M99.
Merging as CHROMIUM patch for now. It will be sent to upstream later,
tracked under b:214038944.

BUG=b:217731002
TEST=Performed the above test.

Signed-off-by: Manish Mandlik <mmandlik@google.com>
Change-Id: I2b7f789329979102339d7e0717522ba417b63109
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3446922
Tested-by: Manish Mandlik <mmandlik@chromium.org>
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Commit-Queue: Manish Mandlik <mmandlik@chromium.org>