CHROMIUM: iwl7000: mvm: rxmq: refactor mac80211 rx_status setting
There's a bunch of duplicated code in the normal data RX and
the no-data RX paths, refactor that.

Note that we're evidently not implementing the 6 GHz band in
the no-data path correctly, need to fix that in firmware
first. Also, we were setting the NSS differently, keep doing
that, but just override the previous values.

Note also that we used to drop frames with bad rate data, to
simplify that just report rate 0 and continue.

Change-Id: I3ee7a1fe1adf8684c48713dbbdc2cbc60bd24cd1
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: d7cb2d8f468355c6c6ea011f51074536e295b129