nt36xxx: Address define directive for TOUCH_STATE_WORKING
touch_state uses a not-equal-to operator with TOUCH_STATE_WORKING.
But TOUCH_STATE_WORKING is not defined at the header; therefore
causing to be detected as an undeclared identifier.

Import the define directive from the xiaomi longcheer edits properly.

This fixes the following compilation error when compiling using lavender-perf_defconfig:

../drivers/input/touchscreen/nt36xxx/nt36xxx.c:574:45: error: use of undeclared identifier 'TOUCH_STATE_WORKING'
        if ( !bTouchIsAwake || (ts->touch_state != TOUCH_STATE_WORKING) ) {
                                                   ^

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>