fs: Spread the usage of `fscrypt_is_nokey_name()` helper function
Commit [1] has renamed DCACHE_ENCRYPTED_NAME to DCACHE_NOKEY_NAME
and commit [2] has introduced a helper function named fscrypt_is_nokey_name()
which gets called when a no-key dentry is found.

There are still some instances of some drivers using the old DCACHE_ENCRYPTED_NAME
variable. So, we need to shift to the new DCACHE_NOKEY_NAME variable. But since
we have got the new helper function, adapt that function to commit [1] and spread
the usage of it across the drivers that use DCACHE_ENCRYPTED_NAME or
DCACHE_NOKEY_NAME to check if there's a no-key dentry.

[1]: eb2969d863 ("fscrypt: rename DCACHE_ENCRYPTED_NAME to DCACHE_NOKEY_NAME")
[2]: 1227ffc9d7 ("fscrypt: add fscrypt_is_nokey_name()")

Change-Id: I5222050ff603ca2c3b6b2ca51a41d116e4784e74
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>