CHROMIUM: Bluetooth: let HCI_QUALITY_REPORT persist over adapter power cycle
The quality report specifications, including AOSP Bluetooth Quality
Report and Intel Telemetry Event, do not define what happen when
the adapter is turned off and then on. To be consistent among
different specifications and vendors, the quality report feature is
turned off when the adapter is powered off and is turned on when
the adapter is powered on if the feature has been on before power
cycle.

BUG=b:228565358
TEST=After enabling BQR, turn off and then on the Bluetooth adapter.
The BQR feature should still be working.

Signed-off-by: Joseph Hwang <josephsih@chromium.org>
Change-Id: I1c81ebb7e5db5c2659f15847e2705fe4f57c9450
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3583639
Reviewed-by: Archie Pusaka <apusaka@chromium.org>