CHROMIUM: iwl7000: mvm: refactor iwl_mvm_cfg_he_sta()
This function sends the STA_HE_CTXT_CMD, which won't be used in the
new MLD API. Instead, a part of the fields of this command
will be in the new STA cmd. Put the parts that are common to both
commands in functions, which will later be used in sending the new
STA cmd.

Change-Id: I0b8512eb64d8f03ff83879bafe9707f897d5b3c6
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: 1c6150fa80cdd80451d7f9110a49c0bc487ae263