zram: zram_drv: Remove non-atomic `pages_compacted` variable
This was a duplicate variable, originally intended to be used
as an atomic type which is already defined above the duplicate
entry. Thanks to the format specifier warning that was triggered
by GCC, we got to know that the duplicate variable has a mismatched
format specified for logging which actually was reading the format
for another argument as it had never a dedicated format specified
to begin with.

This fixes the following warning by GCC:

../drivers/block/zram/zram_drv.c: In function 'mm_stat_show':
../drivers/block/zram/zram_drv.c:1149:69: warning: format '%llu' expects argument of type 'long long unsigned int', but argument 11 has type 'atomic_long_t' {aka 'atomic64_t'} [-Wformat=]
 1149 |                         "%8llu %8llu %8llu %8lu %8ld %8llu %8lu %8llu %8llu %8llu %8llu\n",
      |                                                                 ~~~~^
      |                                                                     |
      |                                                                     long long unsigned int
......
 1157 |                         pool_stats.pages_compacted,
      |                         ~~~~~~~~~~~~~~~~~~~~~~~~~~
      |                                   |
      |                                   atomic_long_t {aka atomic64_t}

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>