CHROMIUM: iwl7000: mvm: avoid UB shift of snif_queue
For the old TX API we need the tfd_queue_msk, but for the
new TX API we don't need it here because we add it to the
station later. However, for the new API mvm->snif_queue is
set to IWL_MVM_INVALID_QUEUE == 0xffff, so the BIT() here
is undefined behaviour.

Since we don't need the tfd_queue_msk value for the new TX
API at all, simply fill it in only for the old API.

Change-Id: I53744fd7cfb6e146a9393272a2a61852841238d9
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: f6af89d0e1c0e01d42a4e531901c85281068f71b