usb: gadget: f_fs: Remove unused variables
With commit [1], our original intention was to just remove
ipc log string and resolve the uninitialized variable warn
for our benefits. It seems commit [1] made that variable's
use case unused and also the ones who are dependent on it
and the similars.

This removes all of the unused variables detected by
GCC after commit [1].

[1]: https://github.com/Atom-X-Devs/android_kernel_xiaomi_scarlet/commit/ae0c64e0171940ecc72a9ae93f3357a808625841

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>