CHROMIUM: iwl7000: mvm: add an indication that the new MLD API is used
WE can't mix between the new MLD API and the old API.
I.e. - we can't send one of the new cmds and then one of the old ones.
This will cause a FW assert. So we need an indication what API should be
used. We use the new API if:
1. FW supports it
2. We are registered to mac80211 with the new MLD ops
Add an indication which will only be true if both conditions are true.

Change-Id: I0adce36d1783cce23d0e080e3c4a8953db33b515
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
iwl7000-tree: b002a08d2a6eb1dc7007cf73325324bc780652ed