CHROMIUM: media: uvcvideo: Change behaviour of quirk
We have two devices where the privacy switch only works during streamon:

1) Jinlon: The hardware is not properly wired
2) degas:  b/228966629

For 1) we decided to return -EBUSY when the device is streamoff, why?
because we cannot tell the real value and we do not want to lie to the
user. This change is downstream and breaks compliance

For 2) Logitech has decided to return the last value during streamon.
This breaks our UI and lies to the user. But passes compliance

Lets unify the behaviour of both in the kernel.

BUG=b:236421061
TEST=v4l2-compliance on jinlon
TEST=Test switch with camera app

Signed-off-by: Ricardo Ribalda <ribalda@chromium.org>
Change-Id: I6c8bcbf7cb5e9b70efc9a6faf2f42754a2f4109e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3711950
Reviewed-by: Sergey Senozhatsky <senozhatsky@chromium.org>