CHROMIUM: mtk-vcodec: fix data_offset buffer overrun
When we validate that the image size does not exceed the size of the
plane buffer, we are currently neglecting to count the data_offset in
our calculation. Most drivers simple ignore data_offset, but the MTK
vcodec driver has been hacked specifically to support it for ARC++
encoding.

BUG=b:235392268
TEST=Kernel builds and runs on elm
Signed-off-by: Justin Green <greenjustin@google.com>

Change-Id: Ic94e6dcbe89c795b60b1922a301b747846604160
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3727210
Tested-by: Justin Green <greenjustin@google.com>
Reviewed-by: Miguel Casas <mcasas@google.com>
Reviewed-by: Miguel Casas-Sanchez <mcasas@chromium.org>
Commit-Queue: Justin Green <greenjustin@google.com>