CHROMIUM: iwl7000: mvm: cleanup beacon_inject_active during hw restart
beacon_inject_active turns true and false via debugfs when we inject a
beacon. During the time of the beacon injection we can't configure the
FW with a beacon cmd. If we have a hw restart during the beacon injection
then in the recovery flow we will not be able to configure the beacon
cmd to the FW. Fix this by cleaning up this variable after an hw restart.

Change-Id: Ia5a0a0b5d8734f63077ceaac936176a345f8d3d1
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
tested: Greenman, Gregory <gregory.greenman@intel.com>
iwl7000-tree: 69529585bbae268a800daf20c6dfd67d501e42a9