f2fs/sysfs: Eliminate redefined read-only attribute macro
Upstream commit [1] has introduced a read-only attribute macro
for f2fs. Meanwhile, we already had an attribute macro set nodes
as read-only since commit [2]. So, this makes the attribute redefined
and causes a warning from the compiler. Hence, remove the duplicate
attribute and opt for the attribute that upstream added.

This fixes the following warning:

../fs/f2fs/sysfs.c:767:9: warning: 'F2FS_RO_ATTR' macro redefined [-Wmacro-redefined]
        ^
../fs/f2fs/sysfs.c:757:9: note: previous definition is here
        ^

[1]: affb55ba78e5 ("f2fs: introduce sysfs atomic write statistics")
[2]: 75706a7b6f07 ("f2fs/sysfs: Introduce a Read-Only attribute macro")

Fixes: affb55ba78e5 ("f2fs: introduce sysfs atomic write statistics")
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>