Revert "BACKPORT: fuse: fix use after free in fuse_read_interrupt()"
This is likely to cause a splat of call trace. Revert this for now and keep
the extensive tests ongoing. If we figure out the root cause for some userspace
services to spit the call trace by the fuse subsystem, we will come back to this
and revert this for a better fix or might completely rebase and chop off this commit
or else we will keep using this.

This reverts commit 5b25fa650692732fa174aee8c2af6eb474180b4c.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>