BACKPORT: FROMLIST: binder: return pending info for frozen async txns
An async transaction to a frozen process will still be successfully
put in the queue. But this pending async transaction won't be processed
until the target process is unfrozen at an unspecified time in the
future. Pass this important information back to the user space caller
by returning BR_TRANSACTION_PENDING_FROZEN.

Bug: 253913841
Test: freeze process and check the return value from binder driver
Link: https://lore.kernel.org/lkml/20221123201654.589322-1-dualli@chromium.org/
Acked-by: Carlos Llamas <cmllamas@google.com>
Signed-off-by: Li Li <dualli@google.com>
Change-Id: Iae575f5f2c4e5e640607c1bee0a466cd21706fae
[ Tashar02: Backport to k4.19 ]
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>