simple_lmk: Default to stock vmpressure trigger threshold
Vmpressure trigger threshold of >= 80 causes apps to be killed very
frequently and thus apps are failing to be consistently kept in
the background on a 4gb ram device. It is to be noted that vmpressure
is the only trigger for slmk to kill app and reclaim memory based on a
specific configuration of minfree and its timeout.
Therefore, switch back to the stock threshold.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>