nt36xxx: Fix return checks for the driver initialization
There has been a pretty wacky macro check to call the err_driver goto label
for touchscreens for lavender. Not only that, it was still proceeding to go
to err_driver regardless of the condition later on. Fix this dumb logic and
remove the GOTO all together since it's also pointless to use it for just one
return check.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>