Revert "nt36xxx: Simplify gesture mode enablement"
enable_gesture_mode is checked at a nested conditional statement to enable
delay_gesture.

If we remove the initialization of enable_gesture_mode during its declaration
and place it as a fallback for the enable_gesture_mode variable's mode evaluation
after that conditional statement, the mentioned statement doesn't get to check
the value of enable_gesture_mode which actually breaks the condition.

We only considered the enablement of enable_gesture_mode but didn't see whether this
variable was already used for any other condition before. This is flawed. Revert this.

This reverts commit 1fdbd8e3d067fb2e00bd2f91d780b164ea2505f3.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>