kernel: printk: Fix userspace log filter
strstr() and strncmp() were never making any comparisons to reach a conclusion. Meanwhile,
the condition was returning "ret" instead of "len" which killed the function before calling
printk_emit(), resulting in producing no userspace logs at all whatsoever. Fix the return
check and flesh it out as a concrete condition rather than a nested condition and replace
the use of strncmp() with strstr() since it's much simpler and doesn't require character
numbers to track to compare the strings. This also disables all messages having the
initials with "logd".

Fixes: 1b64766 ("kernel: printk: Filter out some userspace logs")
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>