kernel: printk: Silence DM_DEV_STATUS failure logs
DM_DEV_STATUS keeps failing for vroot and system in userspace due to its unavailability
and thus pollutes the dmesg with the annoying logs. This filters out the bothersome logs
using strstr() with a comparison of not NULL which implies that what strstr() returns
won't be empty and will match the string. This provides some sanity for readability.

This silences the following logs:

[    4.195838] init: DM_DEV_STATUS failed for vroot: No such device or address
[    4.196434] init: DM_DEV_STATUS failed for system: No such device or address

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>