fingerprint: fpc: Remove unused `irqf` variable
irqf was previously used for bitwise OR operation: assigning the result to the left
operand and passing that variable. Now that we are doing all of the operations
directly without using a custom variable, irq has no use case anymore. It's just
there with a pointless initialization. Remove the unneeded codes.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>