power: Fix Global and Indian device variant detection algorithm
"is_global_version" has been previously been set to true if device variant
is not "Global" which is "India" and thus this inverses the logic of the
algorithm totally. Fix it alongside the usage of the variable for conditions
Moreover, do not fail probe of smb1351 charger driver for WAYNE and LAVENDER
regardless of the device variants since they support the driver without any
abnormalities.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>