xiaomi: fpc: Hardcode vreg to use devm_regulator_get() check
Every vreg in the FPC driver is marked true. Therefore, the "is_optional"
boolean variable is totally useless. We will never hit the check of
devm_regulator_get_optional(). So, remove it all together and hardcode
the check of devm_regulator_get() to determine if returned value is true
for error or NULL pointer.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>