sched/fair: Rewrite compute_energy() to memoize redundant calculations
Calculating the predicted energy for each CPU candidate is expensive in
large part due to a number of redundant calculations resulting in a total
runtime complexity of O(n**2).

By memoizing the redundant energy computations, the runtime complexity of
calculating energy for all CPU candidates is reduced to O(2n), a
significant performance improvement.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
[ Tashar02: Adapt to old feec() and fix calls to compute_energy() ]
Co-authored-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>